+++ FEweek07/src/App.jsx
@@ -2,10 +2,7 @@ import "./App.css";
 import QuizPage from "./pages/QuizPage";
 
 function App() {
-  return 
-  <>
-    <QuizPage />;
-  </>
+  return <QuizPage />;
   
 }
